+++ apps/docs/src/guide/scripts/snippets/initialising-scripts.ts
@@ -5,7 +5,7 @@ import { arrayify, Provider, ReceiptType, ScriptRequest, Wallet } from 'fuels';
 import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../../../env';
 import { CallTestScript } from '../../../typegend';
 
-const provider = await Provider.create(LOCAL_NETWORK_URL);
+const provider = new Provider(LOCAL_NETWORK_URL);
 const wallet = Wallet.fromPrivateKey(WALLET_PVT_KEY, provider);
 const script = new CallTestScript(wallet);