+++ packages/fuels/src/cli/commands/dev/autoStartFuelCore.ts
@@ -25,9 +25,7 @@ export const autoStartFuelCore = async (config: FuelsConfig) => {
 
     const port = config.fuelCorePort ?? (await getPortPromise({ port: 4000 }));
 
-    const providerUrl = `http://${accessIp}:${port}/v1/graphql`;
-
-    const { cleanup, snapshotDir } = await launchNode({
+    const { cleanup, url, snapshotDir } = await launchNode({
       args: [
         ['--snapshot', config.snapshotDir],
         ['--db-type', 'in-memory'],
@@ -38,13 +36,14 @@ export const autoStartFuelCore = async (config: FuelsConfig) => {
       basePath: config.basePath,
       fuelCorePath: config.fuelCorePath,
       includeInitialState: true,
+      killProcessOnExit: true,
     });
 
     fuelCore = {
       bindIp,
       accessIp,
       port,
-      providerUrl,
+      providerUrl: url,
       snapshotDir,
       killChildProcess: cleanup,
     };