+++ packages/utils/src/utils/toUtf8String.ts
@@ -1,6 +1,5 @@
-import type { BytesLike } from '@fuel-ts/interfaces';
-
 import { arrayify } from './arrayify';
+import type { BytesLike } from './arrayify';
 
 /**
  *  These errors are logged when decoding a UTF-8 string fails.