+++ vitest.setup-files.ts
@@ -1,3 +1,4 @@
 import { setupTestMatchers } from '@fuel-ts/utils/test-utils';
+import { expect } from 'vitest';
 
-setupTestMatchers();
+setupTestMatchers(expect);