+++ internal/forc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.89.5
+
+### Patch Changes
+
+- 5ec254d: chore: updated forc version to `0.66.5`
+
 ## 0.89.4
 
 ### Patch Changes