+++ packages/math/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 0.98.0
+
+### Patch Changes
+
+- 12e801e: chore: dependency pinning and auditing
+- Updated dependencies [4ccc628]
+- Updated dependencies [08a31d8]
+- Updated dependencies [0b4c3ab]
+  - @fuel-ts/errors@0.98.0
+
+## 0.97.2
+
+### Patch Changes
+
+- @fuel-ts/errors@0.97.2
+
 ## 0.97.1
 
 ### Patch Changes