+++ packages/versions/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @fuel-ts/versions
 
+## 0.98.0
+
+### Patch Changes
+
+- 12e801e: chore: dependency pinning and auditing
+- 04ca73e: chore: upgrade `fuel-core` to `0.40.2`
+
+## 0.97.2
+
+### Patch Changes
+
+- 5ec254d: chore: updated forc version to `0.66.5`
+
 ## 0.97.1
 
 ### Patch Changes